+++ sinks/cloudwatch/cloudwatch_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stripe/veneur/v14/samplers"
+	"github.com/stripe/veneur/v14/sinks"
 )
 
 // TestServer wraps an internal httptest.Server and provides a convenience
@@ -102,8 +103,9 @@ func TestFlush(t *testing.T) {
 	}(reqBodyCh, done)
 
 	// Flush the sink
-	_, err = sink.Flush(context.Background(), metrics)
+	flushResult, err := sink.Flush(context.Background(), metrics)
 	assert.NoError(t, err)
+	assert.Equal(t, sinks.MetricFlushResult{MetricsFlushed: 3}, flushResult)
 
 	<-done
 }
@@ -139,8 +141,9 @@ func TestFlushWithStandardUnitTagName(t *testing.T) {
 	}(reqBodyCh, done)
 
 	// Flush the sink
-	_, err = sink.Flush(context.Background(), metrics)
+	flushResult, err := sink.Flush(context.Background(), metrics)
 	assert.NoError(t, err)
+	assert.Equal(t, sinks.MetricFlushResult{MetricsFlushed: 1}, flushResult)
 
 	<-done
 }
@@ -179,8 +182,9 @@ func TestFlushWithStripTags(t *testing.T) {
 	}(reqBodyCh, done)
 
 	// Flush the sink
-	_, err = sink.Flush(context.Background(), metrics)
+	flushResult, err := sink.Flush(context.Background(), metrics)
 	assert.NoError(t, err)
+	assert.Equal(t, sinks.MetricFlushResult{MetricsFlushed: 3}, flushResult)
 
 	<-done
 }
@@ -210,8 +214,9 @@ func TestFlushNoop(t *testing.T) {
 	}(reqBodyCh, done)
 
 	// Flush the sink
-	_, err := sink.Flush(context.Background(), metrics)
+	flushResult, err := sink.Flush(context.Background(), metrics)
 	assert.NoError(t, err)
+	assert.Equal(t, sinks.MetricFlushResult{}, flushResult)
 
 	<-done
 }
@@ -227,7 +232,7 @@ func TestFlushRemoteTimeout(t *testing.T) {
 	defer server.Close()
 
 	// Pass non-empty metrics slice
-	metrics := []samplers.InterMetric{{}}
+	metrics := []samplers.InterMetric{{}, {}, {}}
 
 	// Initialize the sink
 	sink := NewCloudwatchMetricSink(
@@ -243,9 +248,10 @@ func TestFlushRemoteTimeout(t *testing.T) {
 		done <- true
 	}(reqBodyCh, done)
 
-	// Assert the flush failed
-	_, err := sink.Flush(context.Background(), metrics)
+	// Assert the flush failed, and metrics were dropped
+	flushResult, err := sink.Flush(context.Background(), metrics)
 	assert.Error(t, err)
+	assert.Equal(t, sinks.MetricFlushResult{MetricsDropped: 3}, flushResult)
 
 	<-done
 }